Special education technology bibliographic data

Contains bibliographic data of journal articles and dissertations on the use of technology in teaching mathematics to students with disabilities.

Usage

SpecialEduTech

Format

An object of class tbl_df (inherits from tbl, data.frame) with 490 rows and 6 columns.

Examples

SpecialEduTech
#> # A tibble: 490 × 6
#>    reference_type  author                            year title keyword abstract
#>    <chr>           <chr>                            <dbl> <chr> <chr>   <chr>   
#>  1 journal_article Block, G. H.                      1980 Dysc… Arithm… Notes t…
#>  2 thesis          Bukatman, K. L.                   1981 The … locus … This st…
#>  3 journal_article Watkins, M. W., & Webb, C.        1981 Comp… Comput… Results…
#>  4 journal_article Chaffin, J. D.                    1982 Arc-… Comput… The Arc…
#>  5 journal_article Chaffin, J. D., Maxwell, B., & …  1982 ARC-… Electr… This ar…
#>  6 thesis          Golden, C. K.                     1982 The … NA      The pur…
#>  7 journal_article Neal, D.                          1982 A re… tradit… Discuss…
#>  8 thesis          Englebert, B. B.                  1983 A st… microc… The pur…
#>  9 thesis          Foster, K.                        1983 The … comput… The eff…
#> 10 journal_article Pommer, L. T.                     1983 Usin… Comput… The art…
#> # ℹ 480 more rows