Detect Multi-Word Expressions

This function detects multi-word expressions (collocations) of specified sizes that appear at least a specified number of times in the provided tokens.

Usage

detect_multi_words(tokens, size = 2:5, min_count = 2)

Arguments

tokens

A tokens object from the quanteda package.

size

A numeric vector specifying the sizes of the collocations to detect (default: 2:5).

min_count

The minimum number of occurrences for a collocation to be considered (default: 2).

Value

A character vector of detected collocations.

Examples

# \donttest{
  mydata <- TextAnalysisR::SpecialEduTech[seq_len(50), ]

  united_tbl <- TextAnalysisR::unite_cols(
    mydata,
    listed_vars = c("title", "keyword", "abstract")
  )

  tokens <- TextAnalysisR::prep_texts(united_tbl, text_field = "united_texts")

  collocations <- TextAnalysisR::detect_multi_words(tokens, size = 2:3, min_count = 2)
  print(collocations)
#>    [1] "assisted instruction"         "computer assisted"           
#>    [3] "learning disabilities"        "disabled students"           
#>    [5] "school students"              "high school"                 
#>    [7] "elementary school"            "mathematics achievement"     
#>    [9] "special education"            "disabilities mathematics"    
#>   [11] "this study"                   "educable mentally"           
#>   [13] "and practice"                 "with learning"               
#>   [15] "multiplication facts"         "mathematics education"       
#>   [17] "problem solving"              "drill and"                   
#>   [19] "significant differences"      "instructional design"        
#>   [21] "mildly handicapped"           "mentally handicapped"        
#>   [23] "learning disabled"            "students with"               
#>   [25] "teacher assisted"             "effects of"                  
#>   [27] "indicated that"               "experimental groups"         
#>   [29] "of the"                       "curriculum based"            
#>   [31] "control groups"               "disabled children"           
#>   [33] "assisted instructional"       "the effects"                 
#>   [35] "microcomputer assisted"       "it is"                       
#>   [37] "determine whether"            "who had"                     
#>   [39] "learning handicapped"         "mental handicaps"            
#>   [41] "more effective"               "on the"                      
#>   [43] "control group"                "use of"                      
#>   [45] "no significant"               "handicapped children"        
#>   [47] "educable mental"              "computerized drill"          
#>   [49] "addition subtraction"         "highly structured"           
#>   [51] "6th graders"                  "multiplication skills"       
#>   [53] "video game"                   "results indicated"           
#>   [55] "study was"                    "this article"                
#>   [57] "word problem"                 "per week"                    
#>   [59] "to be"                        "mastered during"             
#>   [61] "were found"                   "mathematical ability"        
#>   [63] "minutes per"                  "networking system"           
#>   [65] "education special"            "relationship between"        
#>   [67] "without calculators"          "are discussed"               
#>   [69] "arcade game"                  "social studies"              
#>   [71] "intellectual development"     "3rd 6th"                     
#>   [73] "attitude toward"              "multiple baseline"           
#>   [75] "in the"                       "it was"                      
#>   [77] "effectiveness of"             "study examined"              
#>   [79] "teaching methods"             "following treatment"         
#>   [81] "was used"                     "achievement test"            
#>   [83] "as educable"                  "baseline design"             
#>   [85] "could be"                     "task behavior"               
#>   [87] "basic math"                   "and division"                
#>   [89] "for each"                     "instructional procedure"     
#>   [91] "based on"                     "basal group"                 
#>   [93] "do not"                       "computer games"              
#>   [95] "videodisc group"              "all subjects"                
#>   [97] "subtraction multiplication"   "microcomputer technology"    
#>   [99] "mathematical abilities"       "per day"                     
#>  [101] "mild disabilities"            "computer software"           
#>  [103] "mildly mentally"              "multiplication and"          
#>  [105] "results indicate"             "was also"                    
#>  [107] "classified as"                "intermediate grades"         
#>  [109] "the effectiveness"            "division facts"              
#>  [111] "video arcade"                 "child computers"             
#>  [113] "grouping self"                "iq sex"                      
#>  [115] "literature review"            "post tests"                  
#>  [117] "several important"            "experimental group"          
#>  [119] "secondary educable"           "mild mental"                 
#>  [121] "adolescents classified"       "subjects were"               
#>  [123] "was administered"             "videodisc program"           
#>  [125] "counting strategies"          "higher retention"            
#>  [127] "speed acquisition"            "they worked"                 
#>  [129] "correct responding"           "repeated measures"           
#>  [131] "days per"                     "calculators over"            
#>  [133] "elementary secondary"         "solution strategy"           
#>  [135] "speed accuracy"               "word problems"               
#>  [137] "skill acquisition"            "practice game"               
#>  [139] "problems mastered"            "with specific"               
#>  [141] "but had"                      "grades through"              
#>  [143] "which they"                   "who completed"               
#>  [145] "had not"                      "self selected"               
#>  [147] "teacher intervention"         "mentally retarded"           
#>  [149] "math achievement"             "special humans"              
#>  [151] "old male"                     "randomly assigned"           
#>  [153] "scale iq"                     "single subject"              
#>  [155] "had higher"                   "selected goals"              
#>  [157] "only one"                     "accuracy acquisition"        
#>  [159] "posttest scores"              "procedure intervention"      
#>  [161] "of multiplication"            "external locus"              
#>  [163] "goal condition"               "to provide"                  
#>  [165] "specific learning"            "room programs"               
#>  [167] "study investigated"           "aged children"               
#>  [169] "game format"                  "remedial teaching"           
#>  [171] "over time"                    "days after"                  
#>  [173] "educational technology"       "written response"            
#>  [175] "not achieved"                 "time series"                 
#>  [177] "design across"                "it appears"                  
#>  [179] "maintenance test"             "retarded yr"                 
#>  [181] "such as"                      "calculator on"               
#>  [183] "ss completed"                 "addition but"                
#>  [185] "can provide"                  "task within"                 
#>  [187] "per minute"                   "performance iq"              
#>  [189] "drills practice"              "solving task"                
#>  [191] "secondary education"          "test scores"                 
#>  [193] "significantly more"           "computational performance"   
#>  [195] "simple addition"              "the experimental"            
#>  [197] "immediately following"        "significantly better"        
#>  [199] "arithmetic word"              "was employed"                
#>  [201] "arcade games"                 "graders computer"            
#>  [203] "series design"                "achievement level"           
#>  [205] "game environment"             "disabled adolescents"        
#>  [207] "automaticity in"              "significant effect"          
#>  [209] "differences between"          "basic academic"              
#>  [211] "across subjects"              "results suggest"             
#>  [213] "pretest math"                 "feedback procedures"         
#>  [215] "differences in"               "grouping instructional"      
#>  [217] "indicate that"                "suggest that"                
#>  [219] "pencil and"                   "found between"               
#>  [221] "found when"                   "student classified"          
#>  [223] "delivered feedback"           "children with"               
#>  [225] "instructional methods"        "instruction in"              
#>  [227] "in addition"                  "level sex"                   
#>  [229] "two curricula"                "to determine"                
#>  [231] "can be"                       "retention level"             
#>  [233] "achieved automaticity"        "had significant"             
#>  [235] "selected from"                "information processing"      
#>  [237] "interactive videodisc"        "student attitude"            
#>  [239] "junior high"                  "indicated significant"       
#>  [241] "basic skills"                 "during computer"             
#>  [243] "long division"                "design was"                  
#>  [245] "article is"                   "arithmetic remedial"         
#>  [247] "retarded ss"                  "are available"               
#>  [249] "structured program"           "an average"                  
#>  [251] "educational games"            "elementary aged"             
#>  [253] "resource room"                "language arts"               
#>  [255] "disabled elementary"          "instructional strategy"      
#>  [257] "basal curriculum"             "as an"                       
#>  [259] "there were"                   "the calculator"              
#>  [261] "conductance level"            "both groups"                 
#>  [263] "in grades"                    "completed by"                
#>  [265] "in mathematics"               "handicapped learners"        
#>  [267] "handicapped partner"          "groups were"                 
#>  [269] "each student"                 "practice activities"         
#>  [271] "academic skills"              "group based"                 
#>  [273] "general education"            "analysis of"                 
#>  [275] "traditional instruction"      "either drill"                
#>  [277] "yr old"                       "instructional effectiveness" 
#>  [279] "three teachers"               "are presented"               
#>  [281] "training study"               "hand held"                   
#>  [283] "held calculator"              "grade level"                 
#>  [285] "results also"                 "educational microcomputer"   
#>  [287] "problems given"               "development disorder"        
#>  [289] "single digit"                 "investigation was"           
#>  [291] "attention problems"           "teach multiplication"        
#>  [293] "structured teacher"           "mathematics instruction"     
#>  [295] "addition facts"               "and paper"                   
#>  [297] "guidelines are"               "4th 6th"                     
#>  [299] "arithmetic facts"             "an investigation"            
#>  [301] "problems completed"           "study compared"              
#>  [303] "selected computer"            "self esteem"                 
#>  [305] "students computer"            "children while"              
#>  [307] "improved performance"         "teacher implementation"      
#>  [309] "full scale"                   "half hour"                   
#>  [311] "internal external"            "particularly selective"      
#>  [313] "substantial increases"        "visual cues"                 
#>  [315] "problems while"               "using computer"              
#>  [317] "performance characteristics"  "to work"                     
#>  [319] "practice program"             "attitude inventory"          
#>  [321] "help seeking"                 "loosely structured"          
#>  [323] "group received"               "four groups"                 
#>  [325] "implications for"             "responding oral"             
#>  [327] "selective attention"          "present study"               
#>  [329] "arc ed"                       "remedial education"          
#>  [331] "game formats"                 "the study"                   
#>  [333] "repeated practice"            "schools instructional"       
#>  [335] "following conclusions"        "ss were"                     
#>  [337] "problems when"                "handicapped students"        
#>  [339] "in arithmetic"                "administered as"             
#>  [341] "instructional procedures"     "school aged"                 
#>  [343] "toward school"                "strategies for"              
#>  [345] "conceptual understanding"     "motivational features"       
#>  [347] "yr olds"                      "study teaching"              
#>  [349] "software programs"            "performed significantly"     
#>  [351] "to increase"                  "an interactive"              
#>  [353] "math concepts"                "with mild"                   
#>  [355] "game or"                      "assigned to"                 
#>  [357] "behavior analytic"            "iq full"                     
#>  [359] "the videodisc"                "skin conductance"            
#>  [361] "computation skills"           "correct performance"         
#>  [363] "large group"                  "as well"                     
#>  [365] "based measurement"            "achievement child"           
#>  [367] "was assessed"                 "goal structures"             
#>  [369] "achievement gains"            "by their"                    
#>  [371] "grade children"               "strategy choices"            
#>  [373] "mental retardation"           "proportion word"             
#>  [375] "instructional approaches"     "instructional purposes"      
#>  [377] "during teacher"               "education teaching"          
#>  [379] "facts by"                     "at the"                      
#>  [381] "two groups"                   "mathematics curriculum"      
#>  [383] "instructional technology"     "hour per"                    
#>  [385] "spark computerized"           "in basic"                    
#>  [387] "experimental and"             "apple ii"                    
#>  [389] "artificial intelligence"      "cooperative competitive"     
#>  [391] "questions asked"              "active teaching"             
#>  [393] "for all"                      "developing math"             
#>  [395] "computer networking"          "test was"                    
#>  [397] "was found"                    "ed curriculum"               
#>  [399] "retarded children"            "they did"                    
#>  [401] "their use"                    "of control"                  
#>  [403] "division problems"            "computer use"                
#>  [405] "appears to"                   "to improve"                  
#>  [407] "rather than"                  "for the"                     
#>  [409] "after the"                    "high schools"                
#>  [411] "practice methods"             "that selected"               
#>  [413] "procedures were"              "two way"                     
#>  [415] "had conceptual"               "no change"                   
#>  [417] "by the"                       "group grade"                 
#>  [419] "students in"                  "acquisition of"              
#>  [421] "effect of"                    "humans learning"             
#>  [423] "even after"                   "variables were"              
#>  [425] "were discussed"               "were employed"               
#>  [427] "used to"                      "students were"               
#>  [429] "approach was"                 "was conducted"               
#>  [431] "sample was"                   "math skills"                 
#>  [433] "differences existed"          "innovative videodisc"        
#>  [435] "strategy choice"              "school remedial"             
#>  [437] "order to"                     "drills education"            
#>  [439] "or teacher"                   "purpose of"                  
#>  [441] "the subjects"                 "solving achievement"         
#>  [443] "achievement when"             "conclusions were"            
#>  [445] "phases were"                  "were consistently"           
#>  [447] "were randomly"                "were recorded"               
#>  [449] "were trained"                 "curriculum this"             
#>  [451] "an instructional"             "instruction elementary"      
#>  [453] "math computation"             "higher on"                   
#>  [455] "on either"                    "there was"                   
#>  [457] "was designed"                 "on active"                   
#>  [459] "were administered"            "six students"                
#>  [461] "feedback on"                  "be device"                   
#>  [463] "of problems"                  "computers education"         
#>  [465] "used as"                      "disabilities mathematical"   
#>  [467] "disabled 3rd"                 "disabled improved"           
#>  [469] "drill program"                "non learning"                
#>  [471] "teaching basic"               "working in"                  
#>  [473] "in math"                      "reward on"                   
#>  [475] "math automaticity"            "reading and"                 
#>  [477] "that the"                     "found that"                  
#>  [479] "in order"                     "were compared"               
#>  [481] "whether students"             "compared to"                 
#>  [483] "teaching aids"                "basic multiplication"        
#>  [485] "the control"                  "the basal"                   
#>  [487] "the purpose"                  "basic fraction"              
#>  [489] "mathematics concepts"         "for weeks"                   
#>  [491] "algorithm for"                "math microcomputers"         
#>  [493] "practice task"                "the hand"                    
#>  [495] "recorded on"                  "spent on"                    
#>  [497] "trained on"                   "on long"                     
#>  [499] "conditions for"               "for only"                    
#>  [501] "games as"                     "mathematics problem"         
#>  [503] "nondisabled students"         "formats to"                  
#>  [505] "program was"                  "with partner"                
#>  [507] "the teacher"                  "with coaching"               
#>  [509] "group configuration"          "group performed"             
#>  [511] "available for"                "means for"                   
#>  [513] "reading math"                 "was compared"                
#>  [515] "students who"                 "practice games"              
#>  [517] "disadvantaged by"             "male with"                   
#>  [519] "education elementary"         "and pencil"                  
#>  [521] "paper and"                    "on mathematics"              
#>  [523] "the number"                   "disabilities special"        
#>  [525] "on microcomputer"             "level performance"           
#>  [527] "on performance"               "were selected"               
#>  [529] "severely handicapped"         "software use"                
#>  [531] "worked with"                  "subjects in"                 
#>  [533] "embedding drill"              "computerized mathematics"    
#>  [535] "the data"                     "of microcomputer"            
#>  [537] "for these"                    "response for"                
#>  [539] "arithmetic achievement"       "practice on"                 
#>  [541] "effect on"                    "identical problems"          
#>  [543] "story problems"               "teaching paradigm"           
#>  [545] "calculator with"              "computer delivered"          
#>  [547] "problem selections"           "math facts"                  
#>  [549] "computers with"               "work with"                   
#>  [551] "an innovative"                "in teaching"                 
#>  [553] "the tutor"                    "the relationship"            
#>  [555] "computation and"              "posttest and"                
#>  [557] "week for"                     "on paper"                    
#>  [559] "examined the"                 "describes computer"          
#>  [561] "as function"                  "as having"                   
#>  [563] "change group"                 "noted as"                    
#>  [565] "total group"                  "in terms"                    
#>  [567] "the results"                  "number of"                   
#>  [569] "of this"                      "performance of"              
#>  [571] "designed to"                  "to teach"                    
#>  [573] "adolescents with"             "device that"                 
#>  [575] "indicating that"              "differences were"            
#>  [577] "education students"           "system on"                   
#>  [579] "for both"                     "intervention was"            
#>  [581] "posttest was"                 "and control"                 
#>  [583] "disabled no"                  "students child"              
#>  [585] "and second"                   "mathematics system"          
#>  [587] "students receiving"           "automaticity was"            
#>  [589] "education reading"            "reading education"           
#>  [591] "sex and"                      "appear to"                   
#>  [593] "necessary to"                 "to follow"                   
#>  [595] "retarded students"            "found to"                    
#>  [597] "disabled junior"              "assessed in"                 
#>  [599] "the effect"                   "the following"               
#>  [601] "with calculators"             "administration of"           
#>  [603] "influence of"                 "rates of"                    
#>  [605] "ability students"             "them to"                     
#>  [607] "and normal"                   "normal and"                  
#>  [609] "strategies and"               "computer based"              
#>  [611] "instruction learning"         "instruction than"            
#>  [613] "skills to"                    "of reward"                   
#>  [615] "better in"                    "discussed in"                
#>  [617] "experience in"                "increases in"                
#>  [619] "participated in"              "well in"                     
#>  [621] "deficits in"                  "evaluated in"                
#>  [623] "disabled ss"                  "was significant"             
#>  [625] "with mildly"                  "were matched"                
#>  [627] "prior to"                     "structured computer"         
#>  [629] "students classified"          "expectations of"             
#>  [631] "role of"                      "addition learning"           
#>  [633] "for remedial"                 "programs for"                
#>  [635] "achievement in"               "one of"                      
#>  [637] "to have"                      "to their"                    
#>  [639] "mathematics skills"           "investigation of"            
#>  [641] "variables of"                 "task was"                    
#>  [643] "learning disorders"           "increase in"                 
#>  [645] "for use"                      "software for"                
#>  [647] "locus of"                     "terms of"                    
#>  [649] "with traditional"             "was concluded"               
#>  [651] "for ss"                       "teaching mathematics"        
#>  [653] "focused on"                   "period on"                   
#>  [655] "the influence"                "on task"                     
#>  [657] "normal students"              "learning handicaps"          
#>  [659] "acquisition and"              "grades and"                  
#>  [661] "use with"                     "used in"                     
#>  [663] "of video"                     "instruction high"            
#>  [665] "mathematics disability"       "with more"                   
#>  [667] "employed in"                  "the networking"              
#>  [669] "math instruction"             "the first"                   
#>  [671] "the future"                   "the same"                    
#>  [673] "on their"                     "for learning"                
#>  [675] "to assess"                    "able to"                     
#>  [677] "attributed to"                "for teaching"                
#>  [679] "education teachers"           "dealing with"                
#>  [681] "with efficiency"              "to help"                     
#>  [683] "division for"                 "found for"                   
#>  [685] "to two"                       "of computerized"             
#>  [687] "of correct"                   "and feedback"                
#>  [689] "enrolled in"                  "the training"                
#>  [691] "examined in"                  "understanding of"            
#>  [693] "used for"                     "teachers on"                 
#>  [695] "that computer"                "handicapped and"             
#>  [697] "achievement of"               "first and"                   
#>  [699] "given and"                    "goals and"                   
#>  [701] "and five"                     "and nondisabled"             
#>  [703] "and tutorial"                 "and arousal"                 
#>  [705] "and nonhandicapped"           "and respectively"            
#>  [707] "and transfer"                 "and workbook"                
#>  [709] "and number"                   "whether learning"            
#>  [711] "oral computer"                "performance on"              
#>  [713] "on basic"                     "on student"                  
#>  [715] "computer keyboard"            "in attitude"                 
#>  [717] "computer programs"            "education this"              
#>  [719] "students grades"              "the four"                    
#>  [721] "the use"                      "and implications"            
#>  [723] "and non"                      "with an"                     
#>  [725] "traditional and"              "and after"                   
#>  [727] "disabilities in"              "mathematics performance"     
#>  [729] "during the"                   "of an"                       
#>  [731] "ld students"                  "to attentional"              
#>  [733] "to complete"                  "to develop"                  
#>  [735] "to enhanced"                  "to investigate"              
#>  [737] "to participate"               "to solve"                    
#>  [739] "according to"                 "due to"                      
#>  [741] "failed to"                    "led to"                      
#>  [743] "offered to"                   "programmed to"               
#>  [745] "relation to"                  "suited to"                   
#>  [747] "using the"                    "levels of"                   
#>  [749] "modes of"                     "units of"                    
#>  [751] "of sessions"                  "elementary education"        
#>  [753] "the computational"            "application of"              
#>  [755] "areas of"                     "average of"                  
#>  [757] "needs of"                     "series of"                   
#>  [759] "of learners"                  "of responding"               
#>  [761] "in multiplication"            "selected to"                 
#>  [763] "system to"                    "to educational"              
#>  [765] "and written"                  "and one"                     
#>  [767] "investigated the"             "changes in"                  
#>  [769] "deficit in"                   "difference in"               
#>  [771] "in developing"                "in relation"                 
#>  [773] "instruction games"            "students in the"             
#>  [775] "solving were"                 "and motivation"              
#>  [777] "and without"                  "concepts and"                
#>  [779] "was to"                       "instruction mathematics"     
#>  [781] "determine the"                "while the"                   
#>  [783] "than in"                      "students junior"             
#>  [785] "curriculum was"               "with basic"                  
#>  [787] "of general"                   "of regular"                  
#>  [789] "review of"                    "assess the"                  
#>  [791] "the acquisition"              "education teacher"           
#>  [793] "results of"                   "majority of"                 
#>  [795] "of covariance"                "and problem"                 
#>  [797] "among the"                    "improving the"               
#>  [799] "tests the"                    "weeks the"                   
#>  [801] "effective in"                 "of three"                    
#>  [803] "calculator instruction"       "the comparison"              
#>  [805] "the present"                  "the role"                    
#>  [807] "the sample"                   "to one"                      
#>  [809] "children education"           "test and"                    
#>  [811] "treatment and"                "school math"                 
#>  [813] "by learning"                  "the alp"                     
#>  [815] "the authors"                  "the majority"                
#>  [817] "of information"               "of minutes"                  
#>  [819] "of motivation"                "of new"                      
#>  [821] "of pretest"                   "characteristics of"          
#>  [823] "means of"                     "of computer"                 
#>  [825] "instruction computer"         "disabilities problem"        
#>  [827] "assisted math"                "development of"              
#>  [829] "of automaticity"              "and race"                    
#>  [831] "dyad and"                     "keyboard and"                
#>  [833] "retention and"                "and retention"               
#>  [835] "effects on"                   "grade learning"              
#>  [837] "for this"                     "the coaching"                
#>  [839] "the motivational"             "the regular"                 
#>  [841] "achievement mathematics"      "use on"                      
#>  [843] "instruction with"             "from to"                     
#>  [845] "disabled high"                "instruction is"              
#>  [847] "of goal"                      "training of"                 
#>  [849] "amount of"                    "completion of"               
#>  [851] "consisted of"                 "discussion of"               
#>  [853] "elements of"                  "end of"                      
#>  [855] "function of"                  "investigations of"           
#>  [857] "minimum of"                   "of embedding"                
#>  [859] "of questions"                 "of variance"                 
#>  [861] "percentage of"                "perceptions of"              
#>  [863] "rate of"                      "solving learning"            
#>  [865] "of educable"                  "and help"                    
#>  [867] "baseline and"                 "structured and"              
#>  [869] "the student"                  "skills in"                   
#>  [871] "the computation"              "the intervention"            
#>  [873] "the posttest"                 "of addition"                 
#>  [875] "of the learning"              "explored the"                
#>  [877] "education of learning"        "effects for"                 
#>  [879] "the amount"                   "the author"                  
#>  [881] "the completion"               "the computer's"              
#>  [883] "the distribution"             "the end"                     
#>  [885] "of computers"                 "in reading"                  
#>  [887] "the teachers"                 "technology and"              
#>  [889] "feedback to"                  "have the"                    
#>  [891] "the article"                  "the goal"                    
#>  [893] "the second"                   "the variables"               
#>  [895] "level and"                    "and two"                     
#>  [897] "education learning"           "teacher with"                
#>  [899] "instruction on"               "practice education"          
#>  [901] "problems in"                  "accuracy and"                
#>  [903] "and accuracy"                 "retention the"               
#>  [905] "from the"                     "achievement learning"        
#>  [907] "the six"                      "the treatment"               
#>  [909] "study teaching of"            "to the"                      
#>  [911] "used the"                     "teachers in"                 
#>  [913] "practice education of"        "of using"                    
#>  [915] "disabilities and educable"    "assisted and computer"       
#>  [917] "in an"                        "compared the"                
#>  [919] "students microcomputer"       "and traditional"             
#>  [921] "the students in"              "students learning"           
#>  [923] "the attitude"                 "the mathematical"            
#>  [925] "and educable"                 "level of"                    
#>  [927] "of basic"                     "of two"                      
#>  [929] "solving computer"             "instruction for"             
#>  [931] "drill and practice"           "or in"                       
#>  [933] "study was to"                 "calculators the"             
#>  [935] "presented the"                "facts the"                   
#>  [937] "studies of"                   "of mildly"                   
#>  [939] "of selected"                  "program in"                  
#>  [941] "students are"                 "the subjects the"            
#>  [943] "multiplication and division"  "skills of students"          
#>  [945] "performance for"              "children computer"           
#>  [947] "groups of"                    "teaching of"                 
#>  [949] "on math"                      "in performance"              
#>  [951] "elementary students"          "achievement of students"     
#>  [953] "between the"                  "than the"                    
#>  [955] "when the"                     "of data"                     
#>  [957] "of secondary"                 "that learning"               
#>  [959] "performance with"             "learning of multiplication"  
#>  [961] "achievement was"              "disabled and"                
#>  [963] "group instruction"            "and reading"                 
#>  [965] "the pencil"                   "performance and"             
#>  [967] "practice in"                  "to one of"                   
#>  [969] "and students in"              "used the computer"           
#>  [971] "the teacher to"               "normal and learning"         
#>  [973] "of time"                      "the level"                   
#>  [975] "the two"                      "school students learning"    
#>  [977] "teaching mathematics to"      "software to"                 
#>  [979] "teacher to"                   "to use"                      
#>  [981] "practice and"                 "of learning"                 
#>  [983] "experimental and control"     "of four"                     
#>  [985] "of microcomputers"            "of strategy"                 
#>  [987] "for use with"                 "disabilities computer"       
#>  [989] "and traditional instruction"  "performance of learning"     
#>  [991] "the three"                    "games and"                   
#>  [993] "in learning handicapped"      "automaticity in learning"    
#>  [995] "of group"                     "teaching students"           
#>  [997] "disabled elementary students" "in special"                  
#>  [999] "use in"                       "teachers and"                
#>  [ reached 'max' / getOption("max.print") -- omitted 638 entries ]
# }