Evaluate Optimal Number of Topics

This function performs a search for the optimal number of topics (K) using stm::searchK and visualizes diagnostics, including held-out likelihood, residuals, semantic coherence, and lower bound metrics.

Usage

evaluate_optimal_topic_number(
  dfm_object,
  K_range,
  max.em.its = 75,
  categorical_var = NULL,
  continuous_var = NULL,
  verbose = TRUE,
  ...
)

Arguments

dfm_object

A quanteda document-feature matrix (dfm).

K_range

A numeric vector specifying the range of topics (K) to search over.

max.em.its

Maximum number of EM iterations (default: 75).

categorical_var

An optional character string for a categorical variable in the metadata.

continuous_var

An optional character string for a continuous variable in the metadata.

verbose

Logical; if TRUE, prints progress information (default: TRUE).

...

Further arguments passed to stm::searchK.

Value

A Plotly object showing the diagnostics for the number of topics (K).

Examples

if (interactive()) {
 df <- TextAnalysisR::SpecialEduTech
 dfm_object <- df %>%
   TextAnalysisR::preprocess_texts(text_field = "abstract") %>%
   quanteda::dfm() %>%
   quanteda::dfm_trim(min_termfreq = 1, min_docfreq = 1) %>%
   .[quanteda::ntoken(.) > 0, ]
 TextAnalysisR::evaluate_optimal_topic_number(
   dfm_object = dfm_object,
   K_range = 5:30,
   max.em.its = 75,
   categorical_var = "reference_type",
   continuous_var = "year")
}