Plot Word Frequency Across a Continuous Variable

Visualize the word frequency trends across a continuous variable in the metadata.

Usage

word_frequency_trends(
  dfm_object,
  stm_model,
  continuous_variable,
  selected_terms,
  height = 500,
  width = 1000
)

Arguments

dfm_object

A quanteda document-feature matrix (dfm).

stm_model

An STM model object.

continuous_variable

A continuous variable in the metadata.

selected_terms

A vector of terms to analyze trends for.

height

The height of the resulting Plotly plot, in pixels. Defaults to 500.

width

The width of the resulting Plotly plot, in pixels. Defaults to 1000.

Value

A Plotly object visualizing the word frequency trends across the continuous variable.

Details

This function requires a fitted STM model object and a quanteda dfm object.

Examples

if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  dfm_object <- quanteda::dfm(tokens)
  stm_15 <- TextAnalysisR::stm_15
  TextAnalysisR::word_frequency_trends(dfm_object,
                                    stm_model = stm_15,
                                    continuous_variable = "year",
                                    selected_terms = c("calculator", "computer"),
                                    height = 500,
                                    width = 1000)
}