Examine Highest Per-Term Per-Topic Probabilities

Given a tidy data frame of word-topic probabilities (beta values) from an STM model, this function extracts the top terms for each topic.

Usage

examine_top_terms(beta_td, top_n, ...)

Arguments

beta_td

A tidy data frame from tidytext::tidy(stm_model, matrix = "beta"), where stm_model is a fitted Structural Topic Model created using stm::stm().

top_n

The number of top terms per topic to return.

...

Further arguments passed to dplyr::group_by.

Value

A tibble containing the top top_n terms for each topic. The output includes columns for topic, term, and beta values, restricted to the highest-probability terms.

Examples

if (interactive()) {
df <- TextAnalysisR::SpecialEduTech
dfm_object <- df %>%
 preprocess_texts(text_field = "abstract") %>%
 quanteda::dfm()
out <- quanteda::convert(dfm_object, to = "stm")
stm_model <- stm(out$documents,
                out$vocab,
                data = out$meta,
                prevalence = ~
                  I((year >= 1980)*(year - 1980)) +
                  I((year >= 1990)*(year - 1990)) +
                  I((year >= 2000)*(year - 2000)) +
                  I((year >= 2010)*(year - 2010)),
                max.em.its = 75,
                init.type = 'Spectral',
                K = 15,
                verbose = FALSE)

beta_td <- tidytext::tidy(stm_model, matrix="beta")
top_terms <- examine_top_terms(beta_td, top_n = 5)
head(top_terms)
}