Analyze and Visualize Word Correlation Networks

This function creates a word correlation network based on a document-feature matrix (dfm).

Usage

word_correlation_network(
  dfm_object,
  co_occur_n = 30,
  corr_n = 0.4,
  top_node_n = 40,
  height = 1000,
  width = 900
)

Arguments

dfm_object

A quanteda document-feature matrix (dfm).

co_occur_n

Minimum number of co-occurrences for filtering terms (default is 30).

corr_n

Minimum correlation value for filtering terms (default is 0.4).

top_node_n

Number of top nodes to display (default is 40).

height

The height of the resulting Plotly plot, in pixels. Defaults to 1000.

width

The width of the resulting Plotly plot, in pixels. Defaults to 900.

Value

A list containing a Plotly object and a data frame with the results.

Examples

if (interactive()) {
  df <- TextAnalysisR::SpecialEduTech
  united_tbl <- TextAnalysisR::unite_text_cols(df, listed_vars = c("title", "keyword", "abstract"))
  tokens <- TextAnalysisR::preprocess_texts(united_tbl, text_field = "united_texts")
  dfm_object <- quanteda::dfm(tokens)
  word_correlation_network_results <- TextAnalysisR::word_correlation_network(
                                      dfm_object,
                                      co_occur_n = 30,
                                      corr_n = 0.4,
                                      top_node_n = 40,
                                      height = 1000,
                                      width = 900)
  word_correlation_network_results$plot
  word_correlation_network_results$table
  word_correlation_network_results$summary
}