Word Frequency Trends Over Time

Analyze and visualize word frequency trends over time for a fixed term column.

Usage

word_frequency_trends(
  dfm_object,
  gamma_td,
  time_variable,
  selected_terms,
  height = 700,
  width = 1000
)

Arguments

dfm_object

A quanteda dfm object representing document-feature matrix.

gamma_td

A tidy data frame of document-topic probabilities (gamma).

time_variable

The column name for the time variable (e.g., "year").

selected_terms

A vector of terms to analyze trends for.

height

The height of the resulting Plotly plot, in pixels. Defaults to 700.

width

The width of the resulting Plotly plot, in pixels. Defaults to 1000.

Value

A Plotly object showing interactive word frequency trends over time.

Examples

if (interactive()) {
df <- TextAnalysisR::SpecialEduTech
dfm_object <- preprocess_texts(df, text_field = "abstract") %>%
 quanteda::dfm()
out <- quanteda::convert(dfm_object, to = "stm")
stm_15 <- stm::stm(out$documents,
                  out$vocab,
                  data = out$meta,
                  prevalence = ~
                    I((year >= 1980)*(year - 1980)) +
                    I((year >= 1990)*(year - 1990)) +
                    I((year >= 2000)*(year - 2000)) +
                    I((year >= 2010)*(year - 2010)),
                  max.em.its = 75,
                  init.type = 'Spectral',
                  K = 15,
                  verbose = FALSE)
stm_model <- stm_15
gamma_td <- tidytext::tidy(stm_model, matrix = "gamma")
TextAnalysisR::word_frequency_trends(dfm_object,
                                    gamma_td, time_variable = "year",
                                    selected_terms = c("computer", "disability"),
                                    height = 700, width = 1000)
}